+++ src/inspect_ai/_cli/eval.py
@@ -539,13 +539,13 @@ def eval_options(func: Callable[..., Any]) -> Callable[..., click.Context]:
     )
     @click.option(
         "--effort",
-        type=click.Choice(["low", "medium", "high", "max"]),
-        help="Control how many tokens are used for a response, trading off between response thoroughness and token efficiency. Anthropic Claude Opus 4.5 and 4.6 only (`max` only supported on 4.6).",
+        type=click.Choice(["low", "medium", "high", "xhigh", "max"]),
+        help="Control how many tokens are used for a response, trading off between response thoroughness and token efficiency. Anthropic Claude Opus 4.5, 4.6, 4.7 only (`max` only supported on 4.6, `xhigh` only supported on 4.7).",
         envvar="INSPECT_EVAL_EFFORT",
     )
     @click.option(
         "--reasoning-effort",
-        type=click.Choice(["none", "minimal", "low", "medium", "high", "xhigh"]),
+        type=click.Choice(["none", "minimal", "low", "medium", "high", "xhigh", "max"]),
         help="Constrains effort on reasoning. Defaults vary by provider and model and not all models support all values (please consult provider documentation for details).",
         envvar="INSPECT_EVAL_REASONING_EFFORT",
     )