+++ src/inspect_ai/model/_providers/providers.py
@@ -337,7 +337,7 @@ def validate_openai_client(feature: str) -> None:
 
 def validate_anthropic_client(feature: str) -> None:
     PACKAGE = "anthropic"
-    MIN_VERSION = "0.80.0"
+    MIN_VERSION = "0.96.0"
 
     # verify we have the package
     try: